Exercise: added popup, spinner and loading functionality

+++ conFusion/www/js/controllers.js
@@ -67,25 +67,50 @@ angular.module('conFusion.controllers', [])
         };
 })
 
-        .controller('FavoritesController',['$scope','favoriteFactory','menuFactory','baseURL','$ionicListDelegate',function($scope, favoriteFactory, menuFactory, baseURL, $ionicListDelegate){
+        .controller('FavoritesController',['$scope','favoriteFactory','menuFactory','baseURL','$ionicListDelegate','$ionicPopup','$ionicLoading','$timeout',function($scope, favoriteFactory, menuFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout){
 
             $scope.baseURL = baseURL;
             $scope.shouldShowDelete = false;
+
+            $ionicLoading.show({
+                template:'<ion-spinner></ion-spinner>Loading...'
+            });
+
             $scope.favorites = favoriteFactory.getFavorites();
 
             menuFactory.getDishes().query(
                 function(response) {
                     $scope.dishes = response;
-                    $scope.showMenu = true;
+                    //$scope.showMenu = true;
+                    $timeout(function(){
+                        $ionicLoading.hide();
+                    },1000);
                 },
                 function(response) {
                     $scope.message = "Error: "+response.status + " " + response.statusText;
+                     $timeout(function(){
+                        $ionicLoading.hide();
+                    },1000);
                 });
             console.log($scope.dishes, $scope.favorites);
 
             $scope.deleteFavorite = function(index){
 
-                favoriteFactory.deleteFromFavorites(index);
+                var confirmPopup = $ionicPopup.confirm({
+                    title:'Confirm Delete',
+                    template: 'Are you sure you want to delete this item?'
+                });
+
+                confirmPopup.then(function(res){
+                    if(res){
+                        favoriteFactory.deleteFromFavorites(index);
+                        console.log('confirmed delete');
+                    }
+                    else{
+                        console.log('Canceled delete');
+                    }
+                });
+
                 $scope.shouldShowDelete = false;
 
             };
